create BeneficialOwnerDTO
